+++ utils/writeable/src/testing.rs
@@ -50,7 +50,7 @@ impl PartsWrite for TestWriter {
 pub fn writeable_to_parts_for_test<W: Writeable>(
     writeable: &W,
 ) -> (String, Vec<(usize, usize, Part)>) {
-    let mut writer = helpers::TestWriter {
+    let mut writer = TestWriter {
         string: alloc::string::String::new(),
         parts: Vec::new(),
     };
@@ -66,7 +66,7 @@ pub fn writeable_to_parts_for_test<W: Writeable>(
 pub fn try_writeable_to_parts_for_test<W: TryWriteable>(
     writeable: &W,
 ) -> (String, Vec<(usize, usize, Part)>, Option<W::Error>) {
-    let mut writer = helpers::TestWriter {
+    let mut writer = TestWriter {
         string: alloc::string::String::new(),
         parts: Vec::new(),
     };
@@ -76,31 +76,4 @@ pub fn try_writeable_to_parts_for_test<W: TryWriteable>(
         .expect("String writer infallible");
     let (actual_str, actual_parts) = writer.finish();
     (actual_str, actual_parts, result.err())
-}
-
-pub(crate) struct CoreWriteAsPartsWrite<W: fmt::Write + ?Sized>(pub W);
-
-impl<W: fmt::Write + ?Sized> fmt::Write for CoreWriteAsPartsWrite<W> {
-    #[inline]
-    fn write_str(&mut self, s: &str) -> fmt::Result {
-        self.0.write_str(s)
-    }
-
-    #[inline]
-    fn write_char(&mut self, c: char) -> fmt::Result {
-        self.0.write_char(c)
-    }
-}
-
-impl<W: fmt::Write + ?Sized> PartsWrite for CoreWriteAsPartsWrite<W> {
-    type SubPartsWrite = CoreWriteAsPartsWrite<W>;
-
-    #[inline]
-    fn with_part(
-        &mut self,
-        _part: Part,
-        mut f: impl FnMut(&mut Self::SubPartsWrite) -> fmt::Result,
-    ) -> fmt::Result {
-        f(self)
-    }
 }
